+++ packages/game-core/src/index.ts
@@ -57,7 +57,7 @@ export {
 export { CARS, CIVIC_SI, CORE_ROSTER, SPECIAL_ROSTER, getCar } from './data/cars/index.js';
 export { PARTS, getPart } from './data/parts/index.js';
 export type { CpuDifficulty, GarageResult, GarageState, OwnedCarState, PlayerRecord, PurchaseInstallPlan, PurchaseInstallPreview, Transaction } from './garage.js';
-export { CPU_PRIZES, applyAppearance, applyPassStress, applyTune, buyCar, buyPart, canFit, carUnlockReason, createEmptyGarageState, createGarageState, fitPart, modificationBan, ownedCarIds, ownedVehicles, partList, previewPurchaseAndFit, purchaseAndFitPart, removePart, repairCar, repairCost, resolveBuild, selectCar, settleCpuRace } from './garage.js';
+export { CPU_PRIZES, applyAppearance, applyPassStress, applyTune, buyCar, buyPart, canFit, carUnlockReason, createEmptyGarageState, createGarageState, fitPart, gearboxTuningUnlocked, modificationBan, ownedCarIds, ownedVehicles, partList, previewPurchaseAndFit, purchaseAndFitPart, removePart, repairCar, repairCost, resolveBuild, selectCar, settleCpuRace } from './garage.js';
 export type { Appearance,CustomizationItem,DecalItem,DecalPlacement,DecalSurface,GraphicsId,PaintFinishId,VisualComponents,VisualSlot } from './customization.js';
 export { CUSTOMIZATION_CATALOG,CUSTOMIZATION_SCHEMA_VERSION,DECAL_CATALOG,WHEEL_IDS,catalogForSlot,factoryPaintAppearance,migrateAppearance,stockAppearance,validateAppearance } from './customization.js';